+++ NEWS.md
@@ -1,6 +1,7 @@
 ## ib 0.1.1
-- Better documentation and fixed minor typos. 
+- Better documentation and fix minor typos. 
 - Fix typo in `summary` for negative binomial.
+- Enhance simulation for `glm` by allowing user-defined function.
 
 ## ib 0.1.0
 - First version for submission to CRAN. The package is made available on GitHub ([github.com/SMAC-Group/ib/](https://github.com/SMAC-Group/ib/))